third_party/blink/web_tests/platform/generic/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-keyboard-events-expected.txt

Multiple fields UI of datetime-local input type with keyboard events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Please run this with DumpRenderTree.

Test following keys:
Digits
Left/Right - Move focus field inside element
Up/Down - Increment/decrement value of focus field
Tab - Move focus field
Backspace - Make value empty
  
== Digit keys ==
PASS input.value is "98765-09-20T07:56"
== Digit keys starting with zero ==
PASS input.value is "0044-02-03T05:06"
== Digit keys and backspace key ==
PASS input.value is "0008-05-06T09:10"
== Left/Right keys ==
PASS input.value is "0004-09-05T19:05"
PASS document.activeElement.id is "input"
== Up/Down keys ==
PASS input.value is "2012-10-29T00:00"
PASS input.value is "2012-08-29T00:00"
== Up/Down keys on empty value ==
PASS input.value is currentYear + "-11-01T01:59"
== Tab key ==
PASS input.value is "2012-09-05T12:00"
PASS document.activeElement.id is "after"
== Shfit+Tab key ==
PASS input.value is "2012-09-30T13:03"
PASS document.activeElement.id is "before"
== Up key on maximum value ==
PASS input.value is "0001-10-14T00:00"
== Up key with a maximum attribute ==
PASS input.value is "1000-01-01T00:00"
== Down key on minimum value ==
PASS input.value is ""
== Down key with a minimum attribute ==
PASS input.value is ""
== Inconsistent min-max attributes ==
PASS input.value is "1000-12-31T23:59"
PASS input.value is "1999-12-31T23:59"
== Make an invalid date ==
PASS input.value is ""
== Backspace key ==
PASS input.value is ""
== Delete key ==
PASS input.value is ""
== Typeahead ==
PASS input.value is "2012-12-01T00:00"
PASS input.value is "2012-12-02T00:00"
== RTL focus navigation ==
The tests in this block fail on platforms without the lang-attribute-aware-form-control-UI feature.
PASS input.value is "2012-09-01T23:00"
PASS input.value is "2012-02-01T01:00"
PASS input.value is "0008-03-01T01:00"
== Disabled/readonly ==
PASS input.value is "2012-10-08T00:00"
PASS input.value is "2012-11-08T00:00"
PASS input.value is "2012-11-08T00:00"
PASS input.value is "2012-12-08T00:00"
PASS successfullyParsed is true

TEST COMPLETE
